requirements.txt
pymysql~=1.1.1
flask~=3.1.0
python-dotenv~=1.0.1
requests~=2.32.3
sqlalchemy~=2.0.38
tkcalendar~=1.6.1